Simplify resolver registration and class serialization, add example
The server and client should agree on what types are serialized and
what type of markup types need in order to be serialized, because
the types are often shared between them. This moves the resolver
registration responsible for determining what types are serialized
and how they're serialized, to the Configuration class common to
both client and server. Users can override the list of resolvers
to add their own or change the defaults.

It's also nice to be able to just pass classes around without having
to mark them with [DataContract] or [MessagePackObject], so add
a ContractlessResolver to the registration list which handles that.

Change the SimpleServerClient to pass a JobData class around instead
of a string, to demonstrate this.