Make the simple sample ask for the port if it's missing
The simple sample expects a port to be passed into it as a command
line argument. The server will use a dynamic port if it doesn't
have the argument, but the client needs to know which port to
connect to, otherwise it fails with a weird exception.

Make the client sample ask for the port and exit if it doesn't
have a valid one, so it's clear why it fails.